refactor zipping logic to WhyLabsWriter
